Make separate resourcemanagement package
